third_party/blink/web_tests/fast/dom/CSSStyleDeclaration/css-style-declaration-named-setter.html

<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body style="margin: 0px">
<script>

shouldBeEqualToString("document.body.style.margin", "0px");
shouldBe("document.body.style.margin = 1", "1");
shouldBeEqualToString("document.body.style.margin", "1px");

var badString = { toString: function() { throw "Exception in toString()"; } };
shouldThrow("document.body.style.margin = badString", "'Exception in toString()'");
shouldBeEqualToString("document.body.style.margin", "1px"); // Should not reset the previous value.

// Creating a new property on prototype shadows the named property.
shouldThrow("document.body.style.__proto__.margin = 2");
shouldThrow("document.body.style.__proto__.margin");
shouldBeEqualToString("document.body.style.margin", "1px");

shouldBe("document.body.style.margin = 3", "3");
shouldBeEqualToString("document.body.style.margin", "3px");

</script>
</body>
</html>
